relocated placeOrder checkout  function to ReviewDetails component
